Robot vehicle Arduino full: Track location
Update the vehicle location every time it passes an intersection. This
is useful when we are going to a certain waypoint that is more than one
intersection away, since it gives us information about where the vehicle
thinks it is. This is also added in an (invalid) RSSI broadcast packet,
which is helpful for all other vehicles as well as the ground station.