Control panel Settings view: Disable lazy loading
Remove the lazy loading so that we immediately load all the stacked
layouts. This was interfering with loading the overridden settings from
JSON on the ground station to be able to easily resend these settings to
the vehicles.

Improve the creation of the widgets and layouts so that they immediately
have parents, removing the weird contentless Qt window that appears
during loading.