Only consider measurements valid when sending ground station packet
We should only consider a measurement sweep to be valid once we are
sending a ground station packet, not at any other given point, since
then the ground station packet might invalidate once we start moving.
Thus we inline the old `is_measurement_valid` into this case on the
valid callback, and use a flag to determine whether all of the vehicle's
locations are valid.