Waypoint types: Add pass waypoint
This (re)introduces the possibility to add passable waypoints to an
automated mission. This can be useful to require the vehicle to move
around to another waypoint in a certain manner. This we can perform
measurements at waiting locations and then quickly move to faraway
location via a passable waypoint, which is helpful for creating safe
missions during the planning phase, for example.

The `Mission_RF_Sensor` now correctly distinguishes between wait
waypoints that have required sensors, wait waypoints that want to wait
for all other sensors and waypoints that do not want to wait at all.