Control panel Waypoints: Enforce minimal values
The wait ID should be at least 0, since any lower value has no meaning.
The wait count should be at least 1, because a value of zero causes the
range of waypoints to be empty. Provide an error when this is the case
during exporting/sending/compressing, and display invalid columns
visually using a background color. This makes it easier to spot problems
such as values that cannot be casted, or that are too low.