Waypoint types: implement home waypoint
The home waypoint type immediately changes the home location of the
vehicle. It may thus only occur once at the beginning of the waypoints
in the control panel.

Add a method to the `Waypoint` interface to be called after adding the
command, so that the waypoint can update other vehicle state. This is
used to update the home location, and tests show it works as expected.

Other bug fixes in the control panel:
- Remove a method from the waypoints view that was moved to the table
  widget, but its calls were not yet updated.
- Make the waypoint type widget track the item it belongs to, because
  the row (and theoretically the column) could change after the object
  is created, e.g., when we insert or delete a row before this one.
  Create a method to retrieve the item at a certain position, or create
  it if it did not exist already, so that we can ensure that the item
  exists when we create the cell widget.