Implement vehicle pausing
This makes it possible to halt a vehicle (almost immediately) and
continue with the mission later on. We bind this to the pause button of
the infrared sensor, such that it halts the vehicle, stops the RF sensor
from collecting measurements, and idles in the monitor loop. When
pressing the same button again, restore the old mode, start the sensor
and continue the main event loop. Most threads will keep running since
it is too tiresome to deactivate and activate them without resetting the
whole vehicle state.

Also increase the turning speed to improve turning.

Tests are provided for the vehicle mode and its pause handler.