Reduce dependency of Memory_Map on Environment
The Memory_Map only uses a few properties of the Environment, related to
the location of the current vehicle. Create a `Location_Proxy` interface
that ensures that these properties are provided.

Update some Environment methods as well as the tests to work with these
changes, including some simplifications and moving mocks to apply to
other places.