Reconstruction total variation: improve consistency with papers
The factor 0.5 is not necessary as total variation only provides an
additional factor for least squares.

Remove the squaring for beta as it appears to be a typo in the paper
that they correct themselves later on in the paper. Furthermore, beta is
already small enough that it does not need squaring. The derivative that
we used was incorrect because of this, which is also fixed now.