Control panel reconstruction view: implement snapshotting
This patch implements snapshotting the canvas with the reconstructed
image and saving it as a PDF file, making it easy to store reconstructed
images or include them in the thesis, especially since it is high
quality vector image output.

Furthermore disable interpolation to make the snapshot correspond to the
actual view and to make it easier to see exactly which pixels are
assigned a weight higher than zero, as interpolation makes that rather
difficult. Interpolation can still be enabled on request by changing the
parameter in the settings.