Settings: make a base group for iterative reconstructors
Iterative reconstructors (total variation and maximum entropy
regularization to be exact) use an iterative minimization algorithm to
minimize the error. Make a base group for these kinds of algorithms so
we do not have to repeat (and rename) the solver method and solver
iterations.

Furthermore let both algorithms use SLSQP as minimization algorithm with
one iteration. This produces practically the same results as L-BFGS-B
with 20 iterations and is much faster/performant (~118% CPU usage
instead of ~760%). Another advantage is that SLSQP does not need the
`jac` parameter to be defined as it is able to approximate the gradient
(Jacobian) very quickly, contrary to L-BFGS-B which became really slow
when it had to do that. Nevertheless, inspecting the source code of both
algorithms it appears to be faster when it is provided, so we keep
providing it for total variation. We do not provide it for maximum
entropy as it is not trivial to differentiate the entropy function in
such a way that it can be used in a NumPy equation. The main issue when
compared to the total variation factor is that the gradient is applied
to every pixel, whereas the entropy is not (only the distributions of
pixel intensities are used).