Control panel reconstruction view: implement grid view
This view shows the grid under measurement based on the provided size
(both square and rectangular grids are supported). When a measurement
packet arrives, the corresponding link is drawn onto the canvas. By
default all links remain on the canvas, but a checkbox allows you to
toggle that setting to only show the current measurement.

The missions may also be improved because of this view. It allows us to
visualize the coverage of the grid and might help to identify
improvements for the current missions to achieve better coverage in a
shorter time.