Import Manager test: Remove hardcoded package name
The package name is deduced from the root directory name. We assume here
that the root directory is the same name as the repository name, since
that is how the directory is named when cloning. However, one can easily
consider renaming the directory or giving it a different name when
cloning, so we should not make this assumption during the tests.

This alters the tests to deduce the package name from the directory name
in a different way than the code itself does, so we know for certain
that the code does what is expected.