Control panel waypoints view: Reassign imported planning JSON
Make it possible to run the greedy assignment and collision avoidance
algorithms on imported JSON data when it is formatted as output from the
planning runner. This is useful for longer runs which do not happen in
the control panel itself, and makes it possible to delay the collision
avoidance algorithm check until after the run is done, which saves time
and increases the stability of the planning algorithm. This does come
with the risk that some individual missions are not actually safe. If
this is the case, then an error is displayed, but one can also change
the settings to disable the collision avoidance completely.