Planning greedy assignment: Factor in turning time
Certain vehicles may need to turn on a grid in order to keep track of
where they are going, i.e., when they use a line follower sensor. This
should be factored into the planning phase in two ways. The cost of
turning increases the length of the mission, thus it influences the
objective values that we optimize. Secondly, we should try to reduce the
turns in the final assignment by making the assignment algorithm take
the additional cost into account when greedily choosing the next
sensor/vehicle pair combination.

This implements the initial work to get these two items done. The
calculation of distances is factored into its own method since it is now
more complicated than taking some norms and sums, but we reduce the
calculation cost overall because we can precalculate it by vehicle and
then pass the same values over to the distance matrix that works per
vehicle pair, thus reducing the number of combinations calculated. Also
make use of more member variables instead of passing them all around.

The turning cost can be disabled by changing the relevant setting to 0.
Note that the collision avoidance algorithm may override the "distance",
which now includes the turning cost, with something that does not yet
factor the turning time. The tests ensure that the greedy assignment
functionality works.

The code is not very pretty thus far. We do make use of the line
follower direction values to keep track of the current and expected new
directions of each vehicle. At some point, some of the calculation done
with these enum values could be standardized into one place.