Greedy assignment: Optionally assign as `Waypoint` objects
Refactor the way the greedy assignment and collision avoidance algorithm
communicate with each other: pass the new position to the collision
avoidance instead of the entire assignment, since that contains data
that may be in a form that is not very easy to use for anybody but the
waypoints view in the control panel. The collision avoidance algorithm
could have used this reference to add more (pass) waypoints to the
assignment, but that is too tricky in this way. We can later allow it to
return its midpoints to the greedy assignment for optional use.

Now the list form of the waypoints in the assignment is difficult to use
anywhere else, especially since the indexing can stop functioning when
more items are added to it. We would rather replace it entirely with the
`Waypoint` objects, which can store roughly the same data. This would
require a large revamp of the waypoints view, though. Also, a `Waypoint`
stores the position as a `Location` object, which is unhelpful for the
collision avoidance algorithm. Finally, it takes time to create these
objects, which is not wanted when it is embedded in the evolutionary
multiobjective optimization problem.

Long story short, there are reasons for having a `Waypoint` object, but
not always. Thus add a keyword argument to switch to these. We now add
the final waypoint after the collision avoidance did its thing with the
raw coordinate tuple, as a preliminary to possibly adding more points in
between. `Waypoint` object creation requires an import manager, so pass
that through. Clean up some other documentation as well.