Planning runner: Clean up `Algorithm` import
This is the final location that uses `__dict__` to retrieve an imported
class, and it should use the `Import_Manager` for this instead. The
problem is that the algorithm classes do not share the same name as the
submodule (file) that they are (both) in, which is something the import
manager did not support, i.e., it refuses the situation. There is no
reason to disallow this though, so we can now load a class with
a different name than the module that is relative to another module
(relative to a package).

Also clean up the tests and ensure that one exception has the correct
error message.