Planning runner: Avoid unhandled exception at exit of control panel
When the control panel is closed, then the Python process goes through
some termination steps, such as cleaning up its memory use. If another
thread is still busy, and only sees that it should deactivate later on,
then it might end up in a state where it can no longer use all modules
and functions. This causes it to throw (another) exception, which cannot
be handled anymore (especially since it is inside another thread).

Most of our threads end fairly quickly and do not exhibit this problem,
but the planning runner is an exception since the algorithm may be
midway in an iteration when the program is closed. Short-circuit the end
of the thread when it is deactivated to prevent these problems.

This solves some sporadic "unhandled exception" outputs of the control
panel, but does not solve the occasional segmentation fault. This seems
to be unrelated, but the exact source of this problem remains unclear.