Control panel planning view: Integrate sensor grid
This adds a "Grid" button in the planning view which is only enabled for
feasible solutions after the run is done. This then shows the same grid
as in the reconstruction view, slowly building up the sensor links
according to the order in the assignment. It replaces the normal sensor
position plot during this time, but the user can stop it at any time.

This makes use of the `Waypoint` output of the assignment to more easily
deduce data from waypoints belonging to a vehicle and another with which
it synchronizes. This should still work with future additions such as
other waypoint types, but could be improved if we more easily knew which
waypoint is related to which.

Make the `Grid` support passing the configuration without settings or
buffers. Clean up the planning runner and add a `stop` method that
non-destructively ends the run midway, allowing the solutions to be
used/selected again in this case.