Wrap some direction calculations into enum methods
Turn the line follower direction and related classes into `IntEnum`.
This means that no other should break while allowing us to add methods
to the enum values, such as adding, inverting, deducing the coordinate
axis or sign, as well as other methods that have long been in use in the
robot vehicle or greedy assignment. This does mean that every direction
value should always be an enum value, to ensure that we can call methods
on it. Still this should help in making the greedy assignment turning
cost calculation more sensible.