Let AStar know about directions and turning costs
This makes AStar use directions to find out whether it is faster for the
vehicle to use one path compared to the other by reducing turning time.
AStar thus keeps track of the "optimal" direction at each position,
which allows us to know what the final direction is at the end of the
optimal safe path. This is useful for the greedy assignment, which now
let the collision avoidance algorithm provide more information to its
assignment, including the most relevant waypoints along the path.

First of all, in `AStar`, local variables are replaced with member
variables just so that it is easier to split the large algorithm into
multiple methods and reuse the data in other methods. This does mean
that the program might keep more memory as long as the `AStar` object
lives, but this is fine in our cases.

The neighbor offsets of `Geometry` are paired with neighbor directions,
thus simplifying the work to find in which direction the vehicle is
facing at every point. Clean up the order of neighbor offsets to make
more sense. Note that `AStar` and `Collision_Avoidance` both use yaw
angle directions rather than line follower directions so that they could
work with non-grid geometries as well. Conversion of all values is done
by `Greedy_Assignment`.

Return both the full path and the "trending" path with relevant
waypoints, so that callers can decide afterward. For collision
avoidance, we want the full path, but the greedy assignment wants
a simple set of waypoints. There is initial support for adding passing
points in the assignment, but this does not work with the waypoint-based
synchronization at this point and it is thus disabled by default.

Improve performance of some other operations, clean up methods and add
more documentation in `AStar`.