Control panel Waypoints view: Add wait waypoint ID
This adds a column that allows explicit specification of the wait
waypoint ID that the vehicle should receive at the given waypoint. The
wait waypoint ID is not always the same as the waypoint index, as it
skips non-wait waypoints. The indexes also differ between vehicles when
there are more than two, since we may need to wait for some vehicle and
then for another that did not synchronize with the current vehicle.

Since we use 1-indexed waypoints in this view, display them as such, but
internally we convert them to 0-indexed IDs. Change the resize behavior
of the table columns so that the labels fit, and add contextual action
for finding the wait waypoint that is referenced by the current wait
waypoint, enabling actions contextually as well.

Compression keeps in mind that only ranges of wait waypoint IDs can be
compressed, and decompression results in correct ranges again. We add
validation before sending the data, which gives errors for incorrect
referencing or deadlocks.

Support for the wait waypoint is added to the `Waypoint_Wait` object as
well to reflect this addition, even though the waypoints view does not
use these objects, favoring the exportable lists of waypoints. This is
used in later commits.

Also fix a bug with the implicit "wait for vehicle ID" which was
incorrectly calculated to the current vehicle ID in the case of two
vehicles, because the only callers thus far mixed the IDs up as well.