Mission: Improve wait waypoint ID synchronization
This replaces the earlier assumption that all vehicles have the same
number of waypoints and they each align with when they should
synchronize with the required sensors. This is insufficient when there
are more than two vehicles, since they might not synchronize with all
vehicles, thus the waypoint indexes would not align. Instead let the
missions provide a "wait waypoint ID" which is just an identifier for
other vehicles that then know whether the other vehicle is at the
necessary waypoint. This reduces the "hard" dependency on
vehicle-dependent information such as specific indexes and valid
locations, and removes some assumptions in `Mission_Auto` about the
mission consisting of entirely "wait" waypoints, instead using a better
dictionary-based storage. Ranges of wait waypoints are still supported
and missions should still function as before, with sane fallbacks for
missing data. Tests were updated to reflect the new situations and also
to cover more edge cases.