Grid view: Display current sensor/vehicle position
This updates the grid with the "live" positions of the vehicles. During
a mission, the vehicles may pass some intersections before reaching
their measurement location, and they are shown to do so now. Also during
the planning phase, we can see whether there are "pass" waypoints for
detours and whether they function correctly.

The original idea was to show an image of a robot "car" that is moving
around, but the vehicles could be rovers, bipedalic robots, drones, etc.
we simply display a generic sensor antenna block at the intersections.