Control panel Waypoints: highlight error cells
Select and mark cells as invalid when they cause an error during
exporting, sending or other procedures. This makes it easier to find
which value is likely to be the cause of the problem, although in some
cases, such as the validation constraints of wait waypoints, the actual
problem may lie elsewhere. At least it provides an indication to the
user to fix it.

A special exception specific for waypoint table cells with wrong values
keeps track of the data and is handled afterward. Also ensure that
normal validation only occurs when the cell is being edited, instead of
immediately overwriting the error highlight. Finally, alter the pylint
configuration to allow long lines that raise an exception with something
other than just a formatted string.