Mission RF sensor: Dump received mission for reuse
Once we have a complete mission, write the received waypoint packets as
a JSON dump and reload it when the mission is restarted. The dump is
deleted only when a "waypoint_clear" packet is sent to trigger the start
of a new mission. This makes it possible to reuse the mission multiple
times without having to resend the waypoints every time.

Test whether the dumping as well as other functionality of the mission
works correctly.