Control panel RF sensor sender: Sequential packets
To avoid problems with interference and resulting packet drop in
a crowded network caused by packets meant for some (other) vehicle or
acknowledgements that are only useful for the ground station, make the
RF sensor sender (used by Settings and Waypoints views) send their
packets sequentially, starting with the first vehicle and so on. This
causes the sending to take longer, but causes less packet drops and
allows us to reduce the delays between retries.