Infrared sensor: implement powering off the vehicle
This patch allows us to power off the vehicles using a button (in our
case, the "sleep" button) on the remote control. It is mainly useful so
we do not have to connect with Ethernet and SSH into the device every
time to power it off, which also makes it possible to swap batteries
faster.