Planning: Always add home location into assignment
This adds the home location to the assignment of waypoints, such that
they are shown initially in the grid display of the planning view and
are added to the waypoints list to be sent to the vehicles. This means
that we do not need to set the home location via settings on the
vehicles, but can just tweak the home locations during the planning and
update them through the mission waypoints.