Waypoints: Pass along home direction to vehicles
This makes it possible to define the home direction in a home waypoint,
rather than having to change/send it through settings. The assignment
from planning includes this in its home waypoints, and one can change
the direction through a combo box that replaces the columns related to
wait waypoints. In fact, the home direction is in some cases stored in
a form of union with the otherwise unused wait ID to simplify the JSON
and packet structure; the places that need to use the data know which
ones to use. When the home waypoint is used in the RF sensor mission,
the home direction of the vehicle is updated in unison of its location.
This allows the robot vehicle (full Arduino version) to send a command
to update the Arduino's home state (location + home direction).