TDMA scheduler: enforce that sensors only send inside allocated slots
Currently sensors are allowed to send packets even when their current
time is not inside their allocated slot. Especially when the number of
vehicles grows, this may become problematic as other vehicles may be
confused if they receive packets when they do not expect it, causing
unnecessary shifting of schedules. Prevent this by only sending packets
when the current time is inside the slot allocated by the scheduler.