Control panel: resolve segmentation fault at application exit
This happened 100% of the time when the devices view was opened, the
waypoints view was opened and waypoints were sent and after that the
reconstruction view was opened and used. The problem is that Qt does not
appear to cleanup the timers and/or the progress objects by itself, so
we need to do it manually when the process is complete.